+++ main/main.go
@@ -2,18 +2,49 @@ package main
 
 import (
 	"fmt"
-	"os"
 	"github.com/scream870102/argpas"
+	"os"
 )
 
-//this is for test
+type arguments struct {
+	StringArg  string
+	BoolArg    bool
+	IntArg     int
+	FloatArg   float64
+	TriggerArg argpas.Trigger
+}
+
 func main() {
-	args := os.Args
-	res, err := argpas.ParseArgToBool(args[1], "Test")
-	if err != nil {
+	interfaceTest()
+	stringTest()
+	argTest()
+}
+
+func interfaceTest() {
+	var momoko arguments
+	args := os.Args[1:]
+	if err := argpas.ParseToInterface(&momoko, args); err != nil {
+		fmt.Println(err.Error())
+	}
+	fmt.Println(momoko)
+}
+
+func stringTest(){
+	var yuuka string
+	arg := os.Args[1]
+	if result,err:= argpas.ParseToStr(arg,"StringArg");err!=nil{
 		fmt.Println(err.Error())
 	}else{
-		fmt.Println(res)
+		yuuka = result
 	}
+	fmt.Println(yuuka)
+}
 
+func argTest(){
+	arg:= os.Args[1]
+	if result,err:= argpas.Parse(arg,"haruka");err!=nil{
+		fmt.Println(err.Error())
+	}else{
+		fmt.Println(result)
+	}
 }